Allow users to create rule configurations
Entity metadata for associations is not overridden by Doctrine, so when we try
to fetch the name of the derived class of the rule configuration associated with
the rule we want, it comes up with the parent class, which is not helpful.

For now, we need to get this working, so we'll hardcode the class names.