Update build system and composer.json to do PSR-0 loading of the Tillikum library
We need this in order to pull the library in via composer from other places. It
was made a bit complicated by the fact that we want to skip copying of the
vendor directory if we possibly can, since that can be time-consuming.

As a result of these changes, there are fewer symlinks and differences between a
dev build directory and deployed production.